Fix railties test
Parallel tests run tests in multiple processes. stdout is not thread
safe so we can't predict the order in which messages from the test will
be printed out. The reporter has locking mechanism so we can predict
output of the reporter but we can't predict the ouput from just putsing
from a test file.